requirements.txt
# Core dependencies
autogen-agentchat
autogen-ext[magentic-one,openai]

# Web automation
playwright>=1.40

# Data processing
pandas>=2.1
numpy>=1.24

# Visualization
plotly>=5.18
folium
matplotlib

# Document generation
fpdf2
jinja2
markdown

# Database
SQLAlchemy>=2.0.38
alembic==1.13.1
aiosqlite
psycopg2-binary>=2.9.9  # PostgreSQL adapter
asyncpg>=0.29.0  # Async PostgreSQL

# Caching
redis>=5.0
aioredis>=2.0  # Async Redis client

# API
fastapi>=0.100
uvicorn[standard]
python-dotenv==1.0.1
httpx>=0.27.0  # HTTP client for testing
amadeus>=9.0.0  # Amadeus SDK

# Testing
pytest==8.0.2
pytest-asyncio
pytest-cov

# Development tools
black
pylint
mypy
pre-commit 